examples/rag-truehorizon/requirements.txt
openai
faiss-cpu
beautifulsoup4
requests
streamlit
numpy
agno
